Bug: Deleting songs doesnt update queue
