+++ lib/routes/views/Home.dart
@@ -246,7 +246,7 @@ class _HomeState extends State<Home> {
                                     'Delete ${song.title}',
                                     'Are you sure you want to delete ${song.title} by ${song.artist}?',
                                   )) {
-                                    BlocProvider.of<DataBloc>(context)
+                                    BlocProvider.of<QueueBloc>(context)
                                         .add(DeleteSong(song));
                                   }
                                 },
